user_management/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.1.0
dparse==0.5.1
idna==3.3
mongoengine==0.23.1
mypy==0.910
mypy-extensions==0.4.3
packaging==21.3
psycopg2==2.9.2
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
safety==1.10.3
sqlparse==0.2.4
toml==0.10.2
typing-extensions==4.0.0
urllib3==1.26.7
